[Refactoring ⚙️] props대신 custom hook사용 / 기술스택 카테고리 중복 필터 적용
