+++ src/pages/ProjectDetailPage/components/Comments/CommentsItem/ReplyComment.tsx
@@ -1,57 +1,34 @@
+import { UseFormRegisterReturn } from "react-hook-form"
+
 import { Box, Flex } from "@chakra-ui/react"
 import { Comment } from "api-models"
 
-import { CommentsItemProps } from "../../../types/commentItem"
 import CommentsItem from "./CommentsItem"
 
-interface ReplyCommentProps extends Omit<CommentsItemProps, "comment"> {
+interface ReplyCommentProps {
   comment: Comment[]
+  register: UseFormRegisterReturn
 }
 
-const ReplyComment = ({
-  comment,
-  handleOnEdit,
-  handleOffEdit,
-  handleOnReply,
-  handleOffReply,
-  editTargetCommentId,
-  replyTargetCommentId,
-  isReply,
-  isEditing,
-  handleNavigateProfile,
-  handleDelete,
-  handleSubmit,
-  onSubmitEdit,
-  register,
-}: ReplyCommentProps) => {
-  return comment.map((reply) => (
-    <Flex
-      pl="3rem"
-      gap="2rem"
-      w="100%"
-      key={reply.id}>
-      <Box w="100%">
-        <CommentsItem
-          comment={reply}
-          {...{
-            handleOnEdit,
-            handleOffEdit,
-            handleOnReply,
-            handleOffReply,
-            editTargetCommentId,
-            replyTargetCommentId,
-            isReply,
-            isEditing,
-            handleDelete,
-            handleSubmit,
-            onSubmitEdit,
-            register,
-            handleNavigateProfile,
-          }}
-        />
-      </Box>
-    </Flex>
-  ))
+const ReplyComment = ({ comment, register }: ReplyCommentProps) => {
+  return comment.map((reply) => {
+    return (
+      <Flex
+        pl="3rem"
+        gap="2rem"
+        w="100%"
+        key={reply.id}>
+        <Box w="100%">
+          <CommentsItem
+            comment={reply}
+            {...{
+              register,
+            }}
+          />
+        </Box>
+      </Flex>
+    )
+  })
 }
 
 export default ReplyComment
