+++ src/pages/ProjectDetailPage/utils/techStacksCategory.ts
@@ -1,7 +1,7 @@
 import { TechStack } from "api-models"
 
 const techStacksCategory = (techStacks: TechStack[]) => {
-  return Object.entries(
+  const groupedByCateegoryTechStacks = Object.entries(
     techStacks.reduce<Record<string, TechStack[]>>((acc, item) => {
       if (!acc[item.category]) {
         acc[item.category] = []
@@ -10,6 +10,23 @@ const techStacksCategory = (techStacks: TechStack[]) => {
       return acc
     }, {}),
   )
+  console.log(groupedByCateegoryTechStacks)
+
+  const duplicateFilteredTechStacks = groupedByCateegoryTechStacks.map(
+    (item) => {
+      return [
+        item[0],
+        item[1].filter((item2, index, callback) => {
+          return (
+            index === callback.findIndex((t) => t.skill.id === item2.skill.id)
+          )
+        }),
+      ]
+    },
+  )
+
+  // 타입오류 왜 나는지 모름..
+  return duplicateFilteredTechStacks as [string, TechStack[]][]
 }
 
 export default techStacksCategory
